replacing unsigned int with size_t and removed the space

+++ styopin.vladislav/l1/Count.cpp
@@ -2,21 +2,20 @@
 #include <stdexcept>
 #include <limits>
 
-
-void counterSignChanges(unsigned int &countSignChanges, int previous_value, int current)
+void counterSignChanges(size_t &countSignChanges, int previous_value, int current)
 {
   if ((previous_value > 0 && current < 0) || (previous_value < 0 && current > 0)) {
-    if (countSignChanges == std::numeric_limits< unsigned int >::max()) {
+    if (countSignChanges == std::numeric_limits< size_t >::max()) {
       throw std::overflow_error("Overflow");
     }
     countSignChanges++;
   }
 }
 
-void counterMin(unsigned int &countMin, int previous, int current, int& checking)
+void counterMin(size_t &countMin, int previous, int current, int& checking)
 {
   if (checking && previous < current && current != 0) {
-    if (countMin == std::numeric_limits< unsigned int >::max()) {
+    if (countMin == std::numeric_limits< size_t >::max()) {
       throw std::overflow_error("Overflow");
     }
     countMin++;
